NOISSUE Fix type casting in rowCount method to ensure proper integer return type

+++ meshmc/launcher/minecraft/gameoptions/GameOptions.cpp
@@ -128,7 +128,7 @@ QVariant GameOptions::data(const QModelIndex& index, int role) const
 
 int GameOptions::rowCount(const QModelIndex&) const
 {
-	return contents.size();
+	return static_cast<int>(contents.size());
 }
 
 int GameOptions::columnCount(const QModelIndex&) const
